# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator",
        "description": "API documentation for news aggregator",
        "contact": {
            "email": "deekshithshetty48@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/article/set-preference": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Set user article preferences",
                "operationId": "55ce6cb67056fda89205c15eee669424",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "source": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "BBC News"
                                        }
                                    },
                                    "category": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "Society"
                                        }
                                    },
                                    "author": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "BBC News"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response with saved user article preference",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user article preference saved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "prefer_source": {
                                                    "type": "string",
                                                    "example": "BBC News"
                                                },
                                                "prefer_categories": {
                                                    "type": "string",
                                                    "example": "Society"
                                                },
                                                "prefer_author": {
                                                    "type": "string",
                                                    "example": "BBC News,The New York Times"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-31T10:47:16.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-31T10:57:18.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data."
                    },
                    "500": {
                        "description": "Failed to save user article preference"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": [
                            "Token"
                        ]
                    }
                ]
            }
        },
        "/api/user/article/get-preference": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get user article preferences",
                "operationId": "12090da22adf5a104cf18bfbcfc1f009",
                "responses": {
                    "200": {
                        "description": "Successful response with fetched user article preference",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "user article preference fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "prefer_source": {
                                                    "type": "string",
                                                    "example": "BBC News"
                                                },
                                                "prefer_categories": {
                                                    "type": "string",
                                                    "example": "Society"
                                                },
                                                "prefer_author": {
                                                    "type": "string",
                                                    "example": "BBC News,The New York Times"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-31T10:47:16.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-31T10:57:18.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User article preference not set."
                    },
                    "500": {
                        "description": "Failed to fetch user article preference"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearer": [
                            "Token"
                        ]
                    }
                ]
            }
        }
    },
...
}